refactoring to use only one search filter class
